Log forwarding requests
Add logging to the proxy package. Using the same logger as the server
passed as a parameter, so logs are kept in sync and buffer flushed on exit.
Note: The forwarding entry preceds the server access log entry as the
latter is only logged on handler return.
Changed zap config to only log one timestamp in UTC.