Set release mode and remove all proxies trust
Set `release` (production) mode.
Set the list of trusted proxies to nil. By default gin is trusting all
proxies and when running in debug mode the following warning is logged
on start "You trusted all proxies, this is NOT safe. We recommend you to
set a value". For more information:
https://pkg.go.dev/github.com/gin-gonic/gin#Engine.SetTrustedProxies