Add test for forwarded request body
http.Request.Body can only be read once. The next time it is read, an empty
body is returned without indication of an error. Sprayproxy does multiple checks
on the request body and for each one the body is read. More checks can be added
in the future. The added test verifies the original request body is preserved
and forwarded to the backend servers.