keep the path for the proxy request
The proxy backend drops the path so if you use a proxy like https:/someURL/api/webhook (like argocd), it will fall to reach the destination as the path is dropped. 

This change adds the path to the proxy request. 
For backends that did not specify a path, its a noop.